test: fix SLOBAC audit findings (renames, assertion strengthening, deduplication)
- descriptions.test.ts: rename 'should use cyan/blue styling...' to
  'should render description text at the bottom by default'; replace
  vacuous toBeTruthy() with explicit line-index position assertion
- navigation.test.ts: strengthen 'should loop navigation when enabled'
  with explicit wrap-target assertions (❯ ◯ Third / ❯ ◯ First);
  add 'should skip default separator and submit correct values with
  plain string choices' (migrated from separators.test.ts)
- search-filtering.test.ts: rename j-key and k-key tests from
  deliverable-fossil names to behavior-focused names
- selection.test.ts: rename 'should detect readline tab-to-spaces
  conversion bug' to behavior-focused name
- separators.test.ts: deleted; unique scenario migrated to navigation.test.ts
