+++ memory-bank/active/activeContext.md
@@ -1,6 +1,12 @@
 # Active Context
 
 **Current Task:** Fix SLOBAC audit findings (test smells)
-**Phase:** PREFLIGHT - COMPLETE
-**What Was Done:** All 6 preflight checks passed. TDD compliant (test-only changes). Convention compliant. No conflicts. All 8 findings mapped to concrete plan steps.
-**Next Step:** Build phase.
+**Phase:** BUILD - COMPLETE
+**What Was Done:** All 7 implementation steps executed. Renamed 3 deliverable-fossil tests, strengthened loop navigation test + descriptions placement test, migrated separator test to navigation.test.ts, deleted separators.test.ts. 109 tests pass; 4 pre-existing failures are unchanged from branch baseline (confirmed via git stash). Quality gate (format + lint + typecheck) clean.
+**Modified files:**
+- `src/__tests__/descriptions.test.ts`
+- `src/__tests__/navigation.test.ts`
+- `src/__tests__/search-filtering.test.ts`
+- `src/__tests__/selection.test.ts`
+- `src/__tests__/separators.test.ts` (deleted)
+**Next Step:** QA phase.
