+++ memory-bank/active/progress.md
@@ -12,6 +12,19 @@ Fix 8 test smells across 4 test files as identified in `slobac/audit.md`: 3 `del
 * Decisions made
     - All changes are contained within the test suite; no production code changes expected
 
+## 2026-05-13 - BUILD - COMPLETE
+
+* Work completed
+    - Step 1: Renamed 3 deliverable-fossil tests in search-filtering.test.ts and selection.test.ts
+    - Step 2: Strengthened loop navigation test with explicit wrap-target assertions
+    - Step 3: Renamed + strengthened descriptions test with position assertion
+    - Step 4: Migrated string-choice+default-separator test to navigation.test.ts; deleted separators.test.ts
+    - Step 5: Full test suite — 109 pass, 4 pre-existing failures (confirmed unchanged from baseline)
+    - Step 6: Quality gate passed (format, lint, typecheck all clean)
+* Decisions made
+    - Pre-existing failures confirmed via git stash; none introduced by changes
+    - No production code changes required; all fixes were test-only
+
 ## 2026-05-13 - PREFLIGHT - COMPLETE
 
 * Work completed
